Change types and add modifier to getter (#296)
* fix: use uint256 for scaledOngoingDebt

fix: add notNull in isTransferable getter
fix: use uint256 in depletionTimeOf

* test: remove undeeded down casting

* fix: use ">" in depletionTimeOf

refactor: use uint256 for snapshot debt

* address last review comments

* test: lower the rps bound range

+++ benchmark/Flow.Gas.t.sol
@@ -89,7 +89,7 @@ contract Flow_Gas_Test is Integration_Test {
         );
 
         // Deposit amount on an incremented stream ID to make stream solvent.
-        deposit(++streamId, flow.uncoveredDebtOf(streamId) + DEPOSIT_AMOUNT_6D);
+        deposit(++streamId, uint128(flow.uncoveredDebtOf(streamId)) + DEPOSIT_AMOUNT_6D);
 
         // {flow.withdraw} (on a solvent stream).
         computeGas(
