+++ tests/fork/Flow.t.sol
@@ -49,20 +49,20 @@ contract Flow_Fork_Test is Fork_Test {
         uint256 actualAggregateAmount;
         UD21x18 actualRatePerSecond;
         uint40 actualSnapshotTime;
-        uint128 actualSnapshotDebt;
+        uint256 actualSnapshotDebt;
         uint128 actualStreamBalance;
         uint256 actualStreamId;
         uint256 actualTokenBalance;
-        uint128 actualTotalDebt;
+        uint256 actualTotalDebt;
         // Expected values.
         uint256 expectedAggregateAmount;
         UD21x18 expectedRatePerSecond;
         uint40 expectedSnapshotTime;
-        uint128 expectedSnapshotDebt;
+        uint256 expectedSnapshotDebt;
         uint128 expectedStreamBalance;
         uint256 expectedStreamId;
         uint256 expectedTokenBalance;
-        uint128 expectedTotalDebt;
+        uint256 expectedTotalDebt;
     }
 
     /*//////////////////////////////////////////////////////////////////////////
@@ -251,9 +251,9 @@ contract Flow_Fork_Test is Fork_Test {
             newRatePerSecond = ud21x18(newRatePerSecond.unwrap() + 1);
         }
 
-        uint128 beforeSnapshotAmount = flow.getSnapshotDebt(streamId);
-        uint128 totalDebt = flow.totalDebtOf(streamId);
-        uint128 ongoingDebt = flow.ongoingDebtOf(streamId);
+        uint256 beforeSnapshotAmount = flow.getSnapshotDebt(streamId);
+        uint256 totalDebt = flow.totalDebtOf(streamId);
+        uint256 ongoingDebt = flow.ongoingDebtOf(streamId);
 
         // Compute the snapshot time that will be stored post withdraw.
         vars.expectedSnapshotTime = getBlockTimestamp();
@@ -443,7 +443,7 @@ contract Flow_Fork_Test is Fork_Test {
         // If the refundable amount less than 1, deposit some funds.
         if (flow.refundableAmountOf(streamId) <= 1) {
             uint128 depositAmount =
-                flow.uncoveredDebtOf(streamId) + getDefaultDepositAmount(flow.getTokenDecimals(streamId));
+                uint128(flow.uncoveredDebtOf(streamId)) + getDefaultDepositAmount(flow.getTokenDecimals(streamId));
             depositOnStream(streamId, depositAmount);
         }
 
@@ -527,8 +527,8 @@ contract Flow_Fork_Test is Fork_Test {
         // Make sure the requirements are respected.
         address sender = flow.getSender(streamId);
         address recipient = flow.getRecipient(streamId);
-        uint128 uncoveredDebt = flow.uncoveredDebtOf(streamId);
-        uint128 expectedTotalDebt;
+        uint256 uncoveredDebt = flow.uncoveredDebtOf(streamId);
+        uint256 expectedTotalDebt;
 
         resetPrank({ msgSender: sender });
 
@@ -583,7 +583,7 @@ contract Flow_Fork_Test is Fork_Test {
         );
 
         uint256 initialTokenBalance = token.balanceOf(address(flow));
-        uint128 totalDebt = flow.totalDebtOf(streamId);
+        uint256 totalDebt = flow.totalDebtOf(streamId);
 
         vars.expectedSnapshotTime =
             withdrawAmount <= flow.getSnapshotDebt(streamId) ? flow.getSnapshotTime(streamId) : getBlockTimestamp();
