+++ tests/integration/fuzz/adjustRatePerSecond.t.sol
@@ -39,7 +39,7 @@ contract AdjustRatePerSecond_Integration_Fuzz_Test is Shared_Integration_Fuzz_Te
         // Simulate the passage of time.
         vm.warp({ newTimestamp: getBlockTimestamp() + timeJump });
 
-        uint128 previousTotalDebt = flow.totalDebtOf(streamId);
+        uint256 previousTotalDebt = flow.totalDebtOf(streamId);
 
         // Expect the relevant error.
         vm.expectRevert(abi.encodeWithSelector(Errors.SablierFlow_StreamPaused.selector, streamId));
@@ -81,7 +81,7 @@ contract AdjustRatePerSecond_Integration_Fuzz_Test is Shared_Integration_Fuzz_Te
         // Simulate the passage of time.
         vm.warp({ newTimestamp: getBlockTimestamp() + timeJump });
 
-        uint128 previousTotalDebt = flow.totalDebtOf(streamId);
+        uint256 previousTotalDebt = flow.totalDebtOf(streamId);
 
         UD21x18 currentRatePerSecond = flow.getRatePerSecond(streamId);
         if (newRatePerSecond.unwrap() == currentRatePerSecond.unwrap()) {
